Basic gait example has been moved
